requirements.txt
aiodns==3.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-plugins==1.1.1.2
colorama==0.4.6
defusedxml==0.7.1
dnspython==2.8.0
eventlet==0.40.3
filelock==3.19.1
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
geoip2==5.1.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
ipwhois==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
maxminddb==2.8.2
multidict==6.6.4
numpy==2.3.3
OTXv2==1.5.12
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
pycares==4.10.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.2
python-whois==0.9.5
pytz==2025.2
reportlab==4.4.4
requests==2.32.4
requests-file==2.1.0
setuptools==80.9.0
shodan==1.31.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.43
tldextract==5.3.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virustotal-py==0.21.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
xlsxwriter==3.2.5
yarl==1.20.1
abuseipdb
